+++ app/src/main/java/com/example/anatomyviewer/main/MainActivity.kt
@@ -2,40 +2,62 @@ package com.example.anatomyviewer.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.view.View
+import android.widget.FrameLayout
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
+import androidx.navigation.NavController
+import androidx.navigation.findNavController
 import com.example.anatomyviewer.R
 import com.example.anatomyviewer.ar.AnatomyViewerFragment
 import com.example.anatomyviewer.ar.interfaces.ArFragmentResetListener
 import com.example.anatomyviewer.databinding.ActivityMainBinding
-import com.google.android.material.snackbar.Snackbar
+import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-    /*
+class MainActivity : AppCompatActivity(), ArFragmentResetListener, OpenArFragmentListener {
+
     private val TAG: String = MainActivity::class.java.toString()
 
     private lateinit var binding: ActivityMainBinding
 
     private lateinit var mainActivityViewModel: MainActivityViewModel
 
-    lateinit var arFragment: AnatomyViewerFragment
+    private val arFragment: AnatomyViewerFragment by lazy { anatomy_viewer_fragment as AnatomyViewerFragment }
+    private val menu: FrameLayout by lazy { start_menu_layout as FrameLayout }
+    private val navController: NavController by lazy { findNavController(R.id.myNavHostFragment)}
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        mainActivityViewModel = ViewModelProvider(this).get(MainActivityViewModel::class.java)
+        setupDataBinding()
+        setupViewModel()
+        setupArWorld()
+        setupMenu()
+    }
 
+    private fun setupDataBinding(){
         binding = DataBindingUtil.setContentView(this,
             R.layout.activity_main
         )
         binding.lifecycleOwner = this
-        binding.mainActivityViewModel = mainActivityViewModel
+    }
 
-        arFragment = supportFragmentManager.findFragmentById(R.id.anatomy_viewer_fragment) as AnatomyViewerFragment
+    private fun setupViewModel(){
+        mainActivityViewModel = ViewModelProvider(this).get(MainActivityViewModel::class.java)
+        binding.mainActivityViewModel = mainActivityViewModel
+    }
+
+    private fun setupArWorld(){
         arFragment.resetListener = this
+    }
+
+    private fun setupMenu(){
+        menu.
+        menu.visibility = View.VISIBLE
+    }
+
+    override fun openArFragmentRequest() {
+       menu.visibility = View.GONE
     }
 
     override fun resetArFragment() {
@@ -47,5 +69,5 @@ class MainActivity : AppCompatActivity() {
         supportFragmentManager.beginTransaction().replace(R.id.anatomy_viewer_fragment, newArFragment).commit()
     }
 
-     */
+
 }
