+++ app/src/main/java/com/example/anatomyviewer/main/OpenArFragmentListener.kt
@@ -1,0 +1,5 @@
+package com.example.anatomyviewer.main
+
+interface OpenArFragmentListener {
+    fun openArFragmentRequest()
+}
