+++ app/src/main/java/com/example/anatomyviewer/main/WelcomeScreenFragment.kt
@@ -1,0 +1,36 @@
+package com.example.anatomyviewer.main
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Button
+import android.widget.TextView
+import androidx.navigation.NavController
+import androidx.navigation.findNavController
+import androidx.navigation.fragment.findNavController
+import com.example.anatomyviewer.R
+
+
+class WelcomeScreenFragment : Fragment() {
+
+    var listener: OpenArFragmentListener? = null
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.welcome_screen, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        view.findViewById<Button>(R.id.button_first).setOnClickListener {
+            listener?.openArFragmentRequest()
+        }
+    }
+
+}
