+++ app/src/main/res/layout/activity_main.xml
@@ -8,24 +8,46 @@
             type="com.example.anatomyviewer.main.MainActivityViewModel" />
     </data>
 
-
-    <androidx.drawerlayout.widget.DrawerLayout
-        android:id="@+id/drawerLayout"
+    <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <LinearLayout
+        <FrameLayout
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-            <fragment
-                android:id="@+id/myNavHostFragment"
-                android:name="androidx.navigation.fragment.NavHostFragment"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                app:defaultNavHost="true"
-                app:navGraph="@navigation/nav_graph" />
-        </LinearLayout>
+            android:layout_height="match_parent">
 
-    </androidx.drawerlayout.widget.DrawerLayout>
+            <fragment
+                android:id="@+id/anatomy_viewer_fragment"
+                android:name="com.example.anatomyviewer.ar.AnatomyViewerFragment"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent" >
+            </fragment>
+        </FrameLayout>
+
+        <FrameLayout
+            android:id="@+id/start_menu_layout"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <androidx.drawerlayout.widget.DrawerLayout
+                android:id="@+id/drawerLayout"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent">
+
+                <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:orientation="vertical">
+                    <fragment
+                        android:id="@+id/myNavHostFragment"
+                        android:name="androidx.navigation.fragment.NavHostFragment"
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent"
+                        app:defaultNavHost="true"
+                        app:navGraph="@navigation/nav_graph" />
+                </LinearLayout>
+            </androidx.drawerlayout.widget.DrawerLayout>
+        </FrameLayout>
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
