+++ app/src/main/res/layout/welcome_screen.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main.WelcomeScreen">
+    android:background="@android:color/white"
+    tools:context=".main.WelcomeScreenFragment">
 
     <TextView
         android:id="@+id/textview_first"
