+++ app/src/main/res/navigation/nav_graph.xml
@@ -5,14 +5,6 @@
 
     <fragment
         android:id="@+id/welcomeScreen"
-        android:name="com.example.anatomyviewer.main.WelcomeScreen"
-        android:label="WelcomeScreen" >
-        <action
-            android:id="@+id/action_welcomeScreen_to_arMainFragment"
-            app:destination="@id/arMainFragment" />
-    </fragment>
-    <activity
-        android:id="@+id/arMainFragment"
-        android:name="com.example.anatomyviewer.ar.ArMainFragment"
-        android:label="ArMainFragment" />
+        android:name="com.example.anatomyviewer.main.WelcomeScreenFragment"
+        android:label="WelcomeScreen" />
 </navigation>
